Put new facet options in a separate file
https://coveord.atlassian.net/browse/JSUI-2357
